requirements.txt
flask
requests
Flask-APScheduler
mysql-connector-python